fix: workflow from multiple sources
