+++ runtime/src/main/java/io/littlehorse/quarkus/runtime/recordable/LHWorkflowConsumerRecordable.java
@@ -6,30 +6,17 @@ import io.quarkus.runtime.annotations.RecordableConstructor;
 
 import jakarta.enterprise.inject.spi.CDI;
 
-public class LHWorkflowConsumerRecordable implements LHWorkflowRecordable {
-
-    private final Class<?> beanClass;
-    private final String wfSpecName;
+public class LHWorkflowConsumerRecordable extends LHWorkflowRecordable {
 
     @RecordableConstructor
     public LHWorkflowConsumerRecordable(Class<?> beanClass, String wfSpecName) {
-        this.beanClass = beanClass;
-        this.wfSpecName = wfSpecName;
-    }
-
-    public Class<?> getBeanClass() {
-        return beanClass;
-    }
-
-    @Override
-    public String getWfSpecName() {
-        return wfSpecName;
+        super(beanClass, wfSpecName);
     }
 
     @Override
     public void buildWorkflowThread(WorkflowThread workflowThread) {
         LHWorkflowConsumer bean =
-                (LHWorkflowConsumer) CDI.current().select(beanClass).get();
+                (LHWorkflowConsumer) CDI.current().select(getBeanClass()).get();
         bean.accept(workflowThread);
     }
 }
